--- Kotlin/algorithms/hello.kt
@@ -1,5 +1,0 @@
-fun main(){
-
-	println("Hello ajay kotlin")
-
-}
